reestructurando

+++ site/database/models/actorEpisode.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
     const ActorEpisode = sequelize.define(
         "ActorEpisode",
         {
-            screentime: DataTypes.FLOAT(3,1),
+            screentime: DataTypes.FLOAT
         },
         {
             tableName: "actorepisode",
